+++ src/Io.Gate.GateApi/Model/FuturesTrade.cs
@@ -38,7 +38,7 @@ namespace Io.Gate.GateApi.Model
         /// <param name="contract">Futures contract.</param>
         /// <param name="size">Trading size.</param>
         /// <param name="price">Trading price.</param>
-        public FuturesTrade(long id = default(long), decimal createTime = default(decimal), string contract = default(string), long size = default(long), string price = default(string))
+        public FuturesTrade(long id = default(long), double createTime = default(double), string contract = default(string), long size = default(long), string price = default(string))
         {
             this.Id = id;
             this.CreateTime = createTime;
@@ -59,7 +59,7 @@ namespace Io.Gate.GateApi.Model
         /// </summary>
         /// <value>Trading time</value>
         [DataMember(Name="create_time", EmitDefaultValue=false)]
-        public decimal CreateTime { get; set; }
+        public double CreateTime { get; set; }
 
         /// <summary>
         /// Futures contract
